// extensions/conditional-delivery/generated/input.ts
export type CartDeliveryOptionsTransformRunInput = { cart: { deliveryGroups: Array<{ cartLines: Array<{ merchandise:
          | { __typename: 'CustomProduct' }
          | { __typename: 'ProductVariant', id: string }
         }>, deliveryAddress: { countryCode: CountryCode | null } | null, deliveryOptions: Array<{ handle: string, title: string | null }> }> }, deliveryCustomization: { config: { jsonValue: unknown } | null } };