// extensions/conditional-delivery/generated/api.ts
export type InputMaybe<T> = Maybe<T>;
// ...
export type CartDeliveryOptionsTransformRunResult = {
  /**
   * The ordered list of operations to apply to the list of
   * [delivery options](https://shopify.dev/docs/apps/build/checkout/delivery-shipping/delivery-options/build-function).
   */
  operations: Array<Operation>;
};
// ...
export type DeliveryOptionHideOperation = {
  /** The handle of the delivery option to hide. */
  deliveryOptionHandle: Scalars['Handle']['input'];
};
// ...
export type DeliveryOptionMoveOperation = {
  /** The handle of the delivery option to move. */
  deliveryOptionHandle: Scalars['Handle']['input'];
  /** The target index within the delivery group to move the delivery option to. */
  index: Scalars['Int']['input'];
};
// ...
export type DeliveryOptionRenameOperation = {
  /** The handle of the delivery option to rename. */
  deliveryOptionHandle: Scalars['Handle']['input'];
  /** The new name for the delivery option. */
  title: Scalars['String']['input'];
};
// ...
export type Operation = {
  /** An operation that hides a delivery option from a list that's offered to customers at checkout. */
  deliveryOptionHide?: InputMaybe<DeliveryOptionHideOperation>;
  /**
   * An operation that sorts a list of delivery options that are offered to customers at checkout.
   *
   * If you reorder shipping delivery options, then you are
   * [prohibited](https://shopify.dev/docs/apps/launch/app-requirements-checklist#prohibited-app-types).
   * from automatically selecting higher-priced delivery alternatives by default. The cheapest shipping delivery option
   * must always be the first option selected.
   */
  deliveryOptionMove?: InputMaybe<DeliveryOptionMoveOperation>;
  /**
   * An operation that renames a delivery option that's offered to customers at checkout.
   *
   * The carrier name is automatically prepended to the delivery option title at checkout when using the
   * `DeliveryOptionRenameOperation` object, and can't be altered or omitted through the API. For example, if the carrier name
   * is **UPS** and the option is **Standard**, then you could change **UPS Standard** to **UPS Standard Shipping**,
   * but you couldn't change **UPS Standard** to **Standard Shipping**.
   */
  deliveryOptionRename?: InputMaybe<DeliveryOptionRenameOperation>;
};